feat: add User type and integrate with Express Request interface
