+++ src/types/custom.ts
@@ -1,0 +1,7 @@
+import { Types } from "mongoose";
+
+export type User = {
+  userId: Types.ObjectId;
+  email: string;
+  role: string;
+}
