--- src/types/express/express.d.ts
@@ -1,13 +1,0 @@
-import { Types } from 'mongoose';
-
-declare global {
-  namespace Express {
-    export interface Request {
-      user?: {
-        userId: Types.ObjectId;
-        email: string;
-        role: string;
-      };
-    }
-  }
-}
