+++ src/types/express/index.d.ts
@@ -1,0 +1,11 @@
+import { User } from '../custom';
+
+export {}
+
+declare global {
+  namespace Express {
+    export interface Request {
+      user?: User;
+    }
+  }
+}
